+++ src/azure-cli/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+
+**SQL**
+
+* Add support for --partner-database to sql db replica create command
+
 2.13.0
 ++++++